Clone of the popular Pop the Lock game made in Unity with C#.
Based on Indie Nugget's Pop the Lock tutorial on Youtube.

To play the game, click screen to start and the bar will move clockwise, click screen again when it lands on the yellow circle to lower the number in center of lock, the bar will then move in the counterclockwise direction.

Once number decreases to zero, the next level will begin. There are special yellow circle stars that add to score in top left. Have fun!